+++ src/main/java/com/andersen/User/Admin.java
@@ -1,6 +1,6 @@
 package com.andersen.User;
 
-import com.andersen.TicketService.Ticket;
+import com.andersen.concert_tickets.Ticket;
 
 public class Admin extends User {
     Admin(int id , String name) {
